pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator-core"
version = "0.0.1"
description = "Modular orchestration engine for AI agent teams"
requires-python = ">=3.10"
dependencies = [
  "typer==0.12.5",
  "pydantic==2.9.2",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
]

[project.scripts]
orchestrator = "orchestrator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
